feat(impl): implementation for #116
